+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/SchemaGeneratorModule.java
@@ -7,18 +7,18 @@ import com.fasterxml.jackson.databind.type.TypeFactory;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
 import com.github.emm035.openapi.schema.generator.annotations.Extension;
 import com.github.emm035.openapi.schema.generator.annotations.RefPrefix;
-import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
-import com.github.emm035.openapi.schema.generator.base.TypeUtils;
+import com.github.emm035.openapi.schema.generator.internal.Internal;
+import com.github.emm035.openapi.schema.generator.internal.TypeUtils;
 import com.github.emm035.openapi.schema.generator.extension.PropertyExtension;
 import com.github.emm035.openapi.schema.generator.extension.SchemaExtension;
-import com.github.emm035.openapi.schema.generator.nested.SubTypeGenerator;
-import com.github.emm035.openapi.schema.generator.visitors.ArrayFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.BooleanFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.MapFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.NumberFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.ObjectFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.StringFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.SubTypedObjectFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.generators.SubTypeGenerator;
+import com.github.emm035.openapi.schema.generator.internal.visitors.ArrayFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.visitors.BooleanFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.visitors.MapFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.visitors.NumberFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.visitors.ObjectFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.visitors.StringFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.visitors.SubTypedObjectFormatVisitor;
 import com.google.common.collect.Maps;
 import com.google.common.graph.GraphBuilder;
 import com.google.common.graph.MutableGraph;
