+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/internal/generators/NestedSchemaGenerator.java
@@ -1,12 +1,12 @@
-package com.github.emm035.openapi.schema.generator.nested;
+package com.github.emm035.openapi.schema.generator.internal.generators;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
-import com.github.emm035.openapi.schema.generator.visitors.SchemaGeneratorVisitorWrapper;
+import com.github.emm035.openapi.schema.generator.internal.Internal;
+import com.github.emm035.openapi.schema.generator.internal.visitors.SchemaGeneratorVisitorWrapper;
 import com.google.inject.Inject;
 
 public class NestedSchemaGenerator {
