+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/internal/visitors/ArrayFormatVisitor.java
@@ -1,4 +1,4 @@
-package com.github.emm035.openapi.schema.generator.visitors;
+package com.github.emm035.openapi.schema.generator.internal.visitors;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonMappingException;
@@ -9,8 +9,8 @@ import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.ArraySchema;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
 import com.github.emm035.openapi.schema.generator.annotations.Extension;
-import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
-import com.github.emm035.openapi.schema.generator.base.Generator;
+import com.github.emm035.openapi.schema.generator.internal.Internal;
+import com.github.emm035.openapi.schema.generator.internal.Generator;
 import com.github.emm035.openapi.schema.generator.extension.SchemaExtension;
 import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
