+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/internal/visitors/SchemaGeneratorVisitorWrapper.java
@@ -1,4 +1,4 @@
-package com.github.emm035.openapi.schema.generator.visitors;
+package com.github.emm035.openapi.schema.generator.internal.visitors;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonMappingException;
@@ -12,18 +12,11 @@ import com.fasterxml.jackson.databind.jsonFormatVisitors.JsonObjectFormatVisitor
 import com.fasterxml.jackson.databind.jsonFormatVisitors.JsonStringFormatVisitor;
 import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.base.Generator;
-import com.github.emm035.openapi.schema.generator.base.Schemas;
-import com.github.emm035.openapi.schema.generator.base.TypeUtils;
-import com.github.emm035.openapi.schema.generator.exceptions.SchemaGeneratorInternalException;
-import com.github.emm035.openapi.schema.generator.nested.SubTypeGenerator;
-import com.github.emm035.openapi.schema.generator.visitors.ArrayFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.BooleanFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.MapFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.NumberFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.ObjectFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.StringFormatVisitor;
-import com.github.emm035.openapi.schema.generator.visitors.SubTypedObjectFormatVisitor;
+import com.github.emm035.openapi.schema.generator.internal.Generator;
+import com.github.emm035.openapi.schema.generator.internal.Schemas;
+import com.github.emm035.openapi.schema.generator.internal.TypeUtils;
+import com.github.emm035.openapi.schema.generator.internal.SchemaGeneratorInternalException;
+import com.github.emm035.openapi.schema.generator.internal.generators.SubTypeGenerator;
 import com.google.inject.Inject;
 
 public class SchemaGeneratorVisitorWrapper
